fix inter type

+++ domain/domain.c
@@ -9,7 +9,7 @@ unsigned short bilinear_interpolation(float x, float y, pnm ims, int c) {
     int j = x;
     float dx = x - j;
     float dy = y - i;
-    unsigned short interpolation = 0;
+    float interpolation = 0;
     if (i >= 0 && j >= 0 && i < rows && j < cols) {
         interpolation = interpolation + (1-dx) * (1-dy) * pnm_get_component(ims, i, j, c);
     }
@@ -22,5 +22,5 @@ unsigned short bilinear_interpolation(float x, float y, pnm ims, int c) {
     if (i+1 >= 0 && j+1 >= 0 && i+1 < rows && j+1 < cols) {
         interpolation = interpolation + dx * dy * pnm_get_component(ims, i+1, j+1, c);
     }
-    return interpolation;
+    return (unsigned short)interpolation;
 }
